doc: use --numeric-owner while deploy rootfs tarball
In our rootfs tarballs, tar will preserve files' user and group names.
When extracting the rootfs archives into a usb disk or sdcard, tar will
firstly use the user name instead of uid.

For example, the owner of /etc/pn1 in the tarball is pn1(uid:500), while
there is also a pn1 user in the machine to extract the tarball(uid:600).
After extracting, the uid of /etc/pn1 is 600, not the correct 500.

Please note that oe-core will not change this behavior, pseudo and other
systems are used to extract it uses the names that match the system
passwd/group.

So, update the documentions to use --numeric-owner on the tarball
extraction.

CQID: WIND00420376

Acked-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Xin Ouyang <Xin.Ouyang@windriver.com>